Correção de bug e triângulo retângulo
Correção de bug que convertia valor double para float e criação de um script que calcula a área de um triângulo retângulo

+++ projetos/area_circulo/src/area_circulo/Area_circulo.java
@@ -12,7 +12,7 @@ public class Area_circulo {
         Scanner input = new Scanner(System.in);
         
         System.out.println("Qual o raio do círculo?");
-        r = input.nextFloat();
+        r = input.nextDouble();
         
         A = ( 3.14 * Math.pow(r,2)); // calcula a área
